python/flask/requirements.txt
annotated-types==0.7.0
bcrypt==4.2.1
blinker==1.9.0
build==1.2.2.post1
click==8.1.8
cpfpy==0.0.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-cors==5.0.1
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
psycopg==3.2.4
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.40
typing_extensions==4.12.2
waitress==3.0.2
Werkzeug==3.1.3
